--- src/test-utils/mockObject.ts
@@ -1,8 +1,0 @@
-const mockObject = <G>(object: G, overrides?: Partial<G>): G => {
-    return {
-        ...object,
-        ...overrides,
-    };
-};
-
-export default mockObject;
