+++ src/test-utils/mocks/index.ts
@@ -1,0 +1,6 @@
+export const mockObject = <G >(object: G, overrides?: Partial<G>): G => {
+    return {
+        ...object,
+        ...overrides,
+    };
+};
